requirements.txt
# PDF Processing
PyPDF2==3.0.1
pdfminer.six==20221105

# NLP & Named Entity Recognition
spacy==3.7.2

# Machine Learning & Classification
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2

# Semantic Search (FIXED VERSIONS)
sentence-transformers==2.3.1
huggingface-hub==0.20.3
faiss-cpu==1.7.4

# Flask API
flask==3.0.0
flask-cors==4.0.0

# Utilities
numpy==1.24.3
pandas==2.1.4
python-dateutil==2.8.2
